Files migration from Kolab v3 to Kolab v4
Test Plan: `./phpunit`

Reviewers: #hkccp_developers, mollekopf

Reviewed By: #hkccp_developers, mollekopf

Subscribers: #hkccp_developers

Differential Revision: https://git.kolab.org/D5092

+++ src/tests/Feature/DataMigrator/KolabTest.php
@@ -2,10 +2,14 @@
 
 namespace Tests\Feature\DataMigrator;
 
+use App\Backends\Storage;
 use App\DataMigrator\Account;
 use App\DataMigrator\Driver\Kolab\Tags as KolabTags;
 use App\DataMigrator\Engine;
 use App\DataMigrator\Queue as MigratorQueue;
+use App\Fs\Item as FsItem;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage as LaravelStorage;
 use Tests\BackendsTrait;
 use Tests\TestCase;
 
@@ -18,6 +22,9 @@ class KolabTest extends TestCase
 {
     use BackendsTrait;
 
+    private static $skipTearDown = false;
+    private static $skipSetUp = false;
+
     /**
      * {@inheritDoc}
      */
@@ -25,7 +32,12 @@ class KolabTest extends TestCase
     {
         parent::setUp();
 
-        MigratorQueue::truncate();
+        if (!self::$skipSetUp) {
+            MigratorQueue::truncate();
+            FsItem::query()->forceDelete();
+        }
+
+        self::$skipSetUp = false;
     }
 
     /**
@@ -33,7 +45,17 @@ class KolabTest extends TestCase
      */
     public function tearDown(): void
     {
-        MigratorQueue::truncate();
+        if (!self::$skipTearDown) {
+            MigratorQueue::truncate();
+            FsItem::query()->forceDelete();
+
+            $disk = LaravelStorage::disk(\config('filesystems.default'));
+            foreach ($disk->listContents('') as $dir) {
+                $disk->deleteDirectory($dir->path());
+            }
+        }
+
+        self::$skipTearDown = false;
 
         parent::tearDown();
     }
@@ -51,6 +73,9 @@ class KolabTest extends TestCase
         // Run the migration
         $migrator = new Engine();
         $migrator->migrate($src, $dst, ['force' => true, 'sync' => true]);
+
+        $imap = $this->getImapClient($dst_imap);
+        $dav = $this->getDavClient($dst_dav);
 
         // Assert the migrated mail
         $messages = $this->imapList($dst_imap, 'INBOX');
@@ -78,13 +103,13 @@ class KolabTest extends TestCase
         $utf7_folder = \mb_convert_encoding('&kość', 'UTF7-IMAP', 'UTF8');
         $this->assertContains($utf7_folder, $mail_folders);
 
-        // Check migration of folder subscription state
+        // Check migration of folder subscription state and ACL
         $subscribed = $this->imapListFolders($dst_imap, true);
         $this->assertNotContains($utf7_folder, $subscribed);
         $this->assertContains('Test2', $subscribed);
+        $this->assertSame('lrswi', implode('', $imap->getACL('Test2')['john@kolab.org']));
 
         // Assert migrated tags
-        $imap = $this->getImapClient($dst_imap);
         $tags = KolabTags::getKolab4Tags($imap);
         $this->assertCount(1, $tags);
         $this->assertSame('tag', $tags[0]['name']);
@@ -99,7 +124,7 @@ class KolabTest extends TestCase
         $this->assertSame('Party', $events['abcdef']->summary);
         $this->assertSame('Meeting', $events['123456']->summary);
 
-        $events = $this->davList($dst_dav, 'Custom Calendar', Engine::TYPE_EVENT);
+        $events = $this->davList($dst_dav, 'Calendar/Custom Calendar', Engine::TYPE_EVENT);
         $events = \collect($events)->keyBy('uid')->all();
         $this->assertCount(1, $events);
         $this->assertSame('Test Summary', $events['aaa-aaa']->summary);
@@ -117,6 +142,53 @@ class KolabTest extends TestCase
         $this->assertCount(2, $tasks);
         $this->assertSame('Task1', $tasks['ccc-ccc']->summary);
         $this->assertSame('Task2', $tasks['ddd-ddd']->summary);
+
+        // Assert migrated ACL/sharees on a DAV folder
+        $folder = $dav->folderInfo("/calendars/user/{$dst->email}/Default"); // 'Calendar' folder
+        $this->assertCount(1, $folder->invites);
+        $this->assertSame('read-write', $folder->invites['mailto:john@kolab.org']['access']);
+
+        // DAV folder properties (color)
+        $this->assertSame('AABB' . sprintf('%2d', date('d')), $folder->color);
+
+        // Assert migrated files
+        $user = $dst_imap->getUser();
+        $folders = $user->fsItems()->where('type', '&', FsItem::TYPE_COLLECTION)
+            ->select('fs_items.*')
+            ->addSelect(DB::raw("(select value from fs_properties where fs_properties.item_id = fs_items.id"
+                . " and fs_properties.key = 'name') as name"))
+            ->get()
+            ->keyBy('name')
+            ->all();
+        $files = $user->fsItems()->whereNot('type', '&', FsItem::TYPE_COLLECTION)
+            ->select('fs_items.*')
+            ->addSelect(DB::raw("(select value from fs_properties where fs_properties.item_id = fs_items.id"
+                . " and fs_properties.key = 'name') as name"))
+            ->get()
+            ->keyBy('name')
+            ->all();
+        $this->assertSame(2, count($folders));
+        $this->assertSame(3, count($files));
+        $this->assertArrayHasKey('A€B', $folders);
+        $this->assertArrayHasKey('Files', $folders);
+        $this->assertTrue($folders['Files']->children->contains($folders['A€B']));
+        $this->assertTrue($folders['Files']->children->contains($files['empty.txt']));
+        $this->assertTrue($folders['Files']->children->contains($files['&kość.odt']));
+        $this->assertTrue($folders['A€B']->children->contains($files['test2.odt']));
+        $this->assertEquals(10000, $files['&kość.odt']->getProperty('size'));
+        $this->assertEquals(0, $files['empty.txt']->getProperty('size'));
+        $this->assertEquals(10000, $files['test2.odt']->getProperty('size'));
+        $this->assertSame('application/vnd.oasis.opendocument.odt', $files['&kość.odt']->getProperty('mimetype'));
+        $this->assertSame('text/plain', $files['empty.txt']->getProperty('mimetype'));
+        $this->assertSame('application/vnd.oasis.opendocument.odt', $files['test2.odt']->getProperty('mimetype'));
+        $this->assertSame('2024-01-10 09:09:09', $files['&kość.odt']->updated_at->toDateTimeString());
+        $file_content = str_repeat('1234567890', 1000);
+        $this->assertSame($file_content, Storage::fileFetch($files['&kość.odt']));
+        $this->assertSame($file_content, Storage::fileFetch($files['test2.odt']));
+        $this->assertSame('', Storage::fileFetch($files['empty.txt']));
+
+        self::$skipTearDown = true;
+        self::$skipSetUp = true;
     }
 
     /**
@@ -141,6 +213,10 @@ class KolabTest extends TestCase
             '/SUMMARY:Party/' => 'SUMMARY:Test'
         ];
         $this->davAppend($src_dav, 'Calendar', ['event/1.ics'], Engine::TYPE_EVENT, $replace);
+        $this->imapEmptyFolder($src_imap, 'Files');
+        $file_content = rtrim(chunk_split(base64_encode('123'), 76, "\r\n"));
+        $replaces = ['/%FILE%/' => $file_content];
+        $this->imapAppend($src_imap, 'Files', 'kolab3/file1.eml', [], '12-Jan-2024 09:09:09 +0000', $replaces);
 
         // Run the migration
         $migrator = new Engine();
@@ -180,6 +256,21 @@ class KolabTest extends TestCase
         $events = \collect($events)->keyBy('uid')->all();
         $this->assertCount(2, $events);
         $this->assertSame('Test', $events['abcdef']->summary);
+
+        // Assert the migrated files
+        $user = $dst_imap->getUser();
+        $files = $user->fsItems()->whereNot('type', '&', FsItem::TYPE_COLLECTION)
+            ->select('fs_items.*')
+            ->addSelect(DB::raw("(select value from fs_properties where fs_properties.item_id = fs_items.id"
+                . " and fs_properties.key = 'name') as name"))
+            ->get()
+            ->keyBy('name')
+            ->all();
+        $this->assertSame(3, count($files));
+        $this->assertEquals(3, $files['&kość.odt']->getProperty('size'));
+        $this->assertSame('application/vnd.oasis.opendocument.odt', $files['&kość.odt']->getProperty('mimetype'));
+        $this->assertSame('2024-01-12 09:09:09', $files['&kość.odt']->updated_at->toDateTimeString());
+        $this->assertSame('123', Storage::fileFetch($files['&kość.odt']));
     }
 
     /**
@@ -195,7 +286,8 @@ class KolabTest extends TestCase
         }
 
         $kolab3_uri = preg_replace('|^[a-z]+://|', 'kolab3://ned%40kolab.org:simple123@', $imap_uri)
-            . '?dav_host=' . preg_replace('|^davs?://|', '', $dav_uri);
+            . '?dav_host=' . preg_replace('|^davs?://|', '', $dav_uri)
+            . '&v4dav=true';
         $kolab4_uri = preg_replace('|^[a-z]+://|', 'kolab4://jack%40kolab.org:simple123@', $imap_uri)
             . '?dav_host=' . preg_replace('|^davs?://|', '', $dav_uri);
 
@@ -220,7 +312,6 @@ class KolabTest extends TestCase
         // Cleanup the account
         $this->initAccount($imap_account);
         $this->initAccount($dav_account);
-        $this->davDeleteFolder($dav_account, 'Custom Calendar', Engine::TYPE_EVENT);
 
         $imap = $this->getImapClient($imap_account);
 
@@ -235,7 +326,7 @@ class KolabTest extends TestCase
 
         // Create a non-mail folder, we'll assert that it was skipped in migration
         $this->imapCreateFolder($imap_account, 'Test');
-        if (!$imap->setMetadata('Test', ['/private/vendor/kolab/folder-type' => 'contact'])) {
+        if (!$imap->setMetadata('Test', ['/private/vendor/kolab/folder-type' => 'journal'])) {
             throw new \Exception("Failed to set metadata");
         }
         if (!$imap->setMetadata('INBOX', ['/private/vendor/kolab/folder-type' => 'mail.inbox'])) {
@@ -247,14 +338,55 @@ class KolabTest extends TestCase
 
         // One more IMAP folder, subscribed
         $this->imapCreateFolder($imap_account, 'Test2', true);
+        $imap->setACL('Test2', 'john@kolab.org', 'lrswi');
 
-        // Insert some other data to migrate
+        // Calendar, Contact, Tasks, Files folders
+        $utf7_folder = \mb_convert_encoding('Files/A€B', 'UTF7-IMAP', 'UTF8');
+        $folders = [
+            'Calendar' => 'event.default',
+            'Calendar/Custom Calendar' => 'event',
+            'Tasks' => 'task',
+            'Contacts' => 'contact.default',
+            'Files' => 'file.default',
+        ];
+        $folders[$utf7_folder] = 'file';
+        foreach ($folders as $name => $type) {
+            $this->imapCreateFolder($imap_account, $name);
+            $this->imapEmptyFolder($imap_account, $name);
+            if (!$imap->setMetadata($name, ['/private/vendor/kolab/folder-type' => $type])) {
+                throw new \Exception("Failed to set metadata");
+            }
+        }
+        $imap->setACL('Calendar', 'john@kolab.org', 'lrswi');
+        if (!$imap->setMetadata('Calendar', ['/shared/vendor/kolab/color' => 'AABB' . sprintf('%2d', date('d'))])) {
+            throw new \Exception("Failed to set metadata");
+        }
+
+        // Insert some files
+        $file_content = str_repeat('1234567890', 1000);
+        $file_content = rtrim(chunk_split(base64_encode($file_content), 76, "\r\n"));
+        $replaces = ['/%FILE%/' => $file_content];
+        $this->imapAppend($imap_account, 'Files', 'kolab3/file1.eml', [], '10-Jan-2024 09:09:09 +0000', $replaces);
+        $this->imapAppend($imap_account, 'Files', 'kolab3/file2.eml', [], '11-Jan-2024 09:09:09 +0000');
+        $replaces['/&amp;kość.odt/'] = 'test2.odt';
+        $replaces['/&ko%C5%9B%C4%87.odt/'] = 'test2.odt';
+        $this->imapAppend($imap_account, $utf7_folder, 'kolab3/file1.eml', [], '12-Jan-2024 09:09:09 +0000', $replaces);
+
+        // Insert some mail to migrate
+        $this->imapEmptyFolder($imap_account, 'INBOX');
+        $this->imapEmptyFolder($imap_account, 'Drafts');
         $this->imapAppend($imap_account, 'INBOX', 'mail/1.eml');
         $this->imapAppend($imap_account, 'INBOX', 'mail/2.eml', ['SEEN']);
         $this->imapAppend($imap_account, 'Drafts', 'mail/3.eml', ['SEEN']);
+
+        // Insert some DAV data to migrate
+        $this->davCreateFolder($dav_account, 'Calendar/Custom Calendar', Engine::TYPE_EVENT);
+        $this->davEmptyFolder($dav_account, 'Calendar', Engine::TYPE_EVENT);
+        $this->davEmptyFolder($dav_account, 'Calendar/Custom Calendar', Engine::TYPE_EVENT);
+        $this->davEmptyFolder($dav_account, 'Contacts', Engine::TYPE_CONTACT);
+        $this->davEmptyFolder($dav_account, 'Tasks', Engine::TYPE_TASK);
         $this->davAppend($dav_account, 'Calendar', ['event/1.ics', 'event/2.ics'], Engine::TYPE_EVENT);
-        $this->davCreateFolder($dav_account, 'Custom Calendar', Engine::TYPE_EVENT);
-        $this->davAppend($dav_account, 'Custom Calendar', ['event/3.ics'], Engine::TYPE_EVENT);
+        $this->davAppend($dav_account, 'Calendar/Custom Calendar', ['event/3.ics'], Engine::TYPE_EVENT);
         $this->davAppend($dav_account, 'Contacts', ['contact/1.vcf', 'contact/2.vcf'], Engine::TYPE_CONTACT);
         $this->davAppend($dav_account, 'Tasks', ['task/1.ics', 'task/2.ics'], Engine::TYPE_TASK);
     }
@@ -266,7 +398,7 @@ class KolabTest extends TestCase
     {
         $this->initAccount($imap_account);
         $this->initAccount($dav_account);
-        $this->davDeleteFolder($dav_account, 'Custom Calendar', Engine::TYPE_EVENT);
+        $this->davDeleteFolder($dav_account, 'Calendar/Custom Calendar', Engine::TYPE_EVENT);
         $this->imapDeleteFolder($imap_account, 'Test');
         $this->imapDeleteFolder($imap_account, 'Test2');
         $this->imapDeleteFolder($imap_account, \mb_convert_encoding('&kość', 'UTF7-IMAP', 'UTF8'));
